Move Assets functionality into ServiceDeskModule
- Assets is part of Jira Service Management, not a separate product
- Move assets_list_workspaces tool into ServiceDeskModule
- Update resource URI to service-desk://assets/workspaces
- Remove separate AssetsModule
- Update module manager to reflect 3 core modules: jira, confluence, service_desk

This better reflects the actual Atlassian product structure.
