+++ src/atlassian_mcp_server/modules/__init__.py
@@ -4,12 +4,10 @@ from .base import BaseModule
 from .jira import JiraModule
 from .confluence import ConfluenceModule
 from .service_desk import ServiceDeskModule
-from .assets import AssetsModule
 
 __all__ = [
     "BaseModule",
     "JiraModule", 
     "ConfluenceModule",
-    "ServiceDeskModule",
-    "AssetsModule"
+    "ServiceDeskModule"
 ]
